requirements.txt
pandas
numpy
requests
upstox-python-sdk
python-telegram-bot
python-dotenv
pyyaml
